fix(core): rid relation between command and event
